Update setup initiate session to use determine notice type
https://eaflood.atlassian.net/browse/WATER-5101

We previously implemented 'ad-hoc' returns as a journey for notices. This is going to change.

We will introduce a page to allow the user to select a notice type when they have chosen a licence for notices.

As part of this upcoming page we need to extract our existing journey logic into a reusable function.

This change introduces the Determine Notice Type Service. This function will set the notice type for all journeys.

Our initial entry to setting up the session data is still based on the query string, but this may change.
