+++ test/services/notices/setup/initiate-session.service.test.js
@@ -24,120 +24,26 @@ describe('Notices - Setup - Initiate Session service', () => {
   })
 
   describe('when called', () => {
+    it('creates a new session record', async () => {
+      const result = await InitiateSessionService.go('invitations')
+
+      const matchingSession = await SessionModel.query().findById(result.sessionId)
+
+      expect(matchingSession.data).to.equal({
+        journey: 'invitations',
+        name: 'Returns: invitation',
+        notificationType: 'Returns invitation',
+        referenceCode: matchingSession.referenceCode, // randomly generated
+        subType: 'returnInvitation'
+      })
+    })
+
     it('correctly returns the redirect path and session id', async () => {
       const result = await InitiateSessionService.go('invitations')
 
       expect(result).to.equal({
         sessionId: result.sessionId,
         path: 'returns-period'
-      })
-    })
-
-    describe('when the "notificationType" is "invitations"', () => {
-      it('creates a new session record', async () => {
-        const result = await InitiateSessionService.go('invitations')
-
-        const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-        expect(matchingSession.data).to.equal({
-          journey: 'invitations',
-          name: 'Returns: invitation',
-          notificationType: 'Returns invitation',
-          referenceCode: matchingSession.referenceCode, // randomly generated
-          subType: 'returnInvitation'
-        })
-      })
-
-      it('correctly returns the redirect path and session id', async () => {
-        const result = await InitiateSessionService.go('invitations')
-
-        expect(result).to.equal({
-          sessionId: result.sessionId,
-          path: 'returns-period'
-        })
-      })
-
-      describe('the "referenceCode" property', () => {
-        it('returns a reference code for "invitations" notifications', async () => {
-          const result = await InitiateSessionService.go('invitations')
-
-          const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-          expect(matchingSession.referenceCode).to.include('RINV-')
-          expect(matchingSession.referenceCode.length).to.equal(11)
-        })
-      })
-    })
-
-    describe('when the "notificationType" is "reminders"', () => {
-      it('creates a new session record', async () => {
-        const result = await InitiateSessionService.go('reminders')
-
-        const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-        expect(matchingSession.data).to.equal({
-          journey: 'reminders',
-          name: 'Returns: reminder',
-          notificationType: 'Returns reminder',
-          referenceCode: matchingSession.referenceCode, // randomly generated
-          subType: 'returnReminder'
-        })
-      })
-
-      it('correctly returns the redirect path and session id', async () => {
-        const result = await InitiateSessionService.go('reminders')
-
-        expect(result).to.equal({
-          sessionId: result.sessionId,
-          path: 'returns-period'
-        })
-      })
-
-      describe('the "referenceCode" property', () => {
-        it('returns a reference code for "reminders" notifications', async () => {
-          const result = await InitiateSessionService.go('reminders')
-
-          const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-          expect(matchingSession.referenceCode).to.include('RREM-')
-          expect(matchingSession.referenceCode.length).to.equal(11)
-        })
-      })
-    })
-
-    describe('when the "notificationType" is "ad-hoc"', () => {
-      it('creates a new session record', async () => {
-        const result = await InitiateSessionService.go('ad-hoc')
-
-        const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-        expect(matchingSession.data).to.equal({
-          journey: 'ad-hoc',
-          name: 'Returns: ad-hoc',
-          notificationType: 'Ad hoc',
-          referenceCode: matchingSession.referenceCode, // randomly generated
-          subType: 'adHocReminder'
-        })
-      })
-
-      it('correctly returns the redirect path and session id', async () => {
-        const result = await InitiateSessionService.go('ad-hoc')
-
-        expect(result).to.equal({
-          sessionId: result.sessionId,
-          path: 'licence'
-        })
-      })
-
-      describe('the "referenceCode" property', () => {
-        it('returns a reference code for an "ad-hoc" notification', async () => {
-          const result = await InitiateSessionService.go('ad-hoc')
-
-          const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-          expect(matchingSession.referenceCode).to.include('ADHC-')
-          expect(matchingSession.referenceCode.length).to.equal(11)
-        })
       })
     })
 
@@ -181,17 +87,6 @@ describe('Notices - Setup - Initiate Session service', () => {
         expect(result).to.equal({
           sessionId: result.sessionId,
           path: 'abstraction-alerts/alert-type'
-        })
-      })
-
-      describe('the "referenceCode" property', () => {
-        it('returns a reference code for an "ad-hoc" notification', async () => {
-          const result = await InitiateSessionService.go('abstraction-alert', monitoringStationId)
-
-          const matchingSession = await SessionModel.query().findById(result.sessionId)
-
-          expect(matchingSession.referenceCode).to.include('WAA-')
-          expect(matchingSession.referenceCode.length).to.equal(10)
         })
       })
     })
